Move upcoming training dates methods to a separate service
And some related methods to some other services too.
