+++ site/app/Form/TrainingApplicationAdminFormFactory.php
@@ -6,6 +6,7 @@ namespace MichalSpacekCz\Form;
 use MichalSpacekCz\Form\Controls\TrainingControlsFactory;
 use MichalSpacekCz\Training\Applications;
 use MichalSpacekCz\Training\Dates\TrainingDates;
+use MichalSpacekCz\Training\Dates\UpcomingTrainingDates;
 use MichalSpacekCz\Training\Statuses;
 use Nette\Application\UI\Form;
 use Nette\Database\Row;
@@ -24,6 +25,7 @@ class TrainingApplicationAdminFormFactory
 		private readonly FormFactory $factory,
 		private readonly Applications $trainingApplications,
 		private readonly TrainingDates $trainingDates,
+		private readonly UpcomingTrainingDates $upcomingTrainingDates,
 		private readonly TrainingControlsFactory $trainingControlsFactory,
 		private readonly Statuses $trainingStatuses,
 	) {
@@ -44,7 +46,7 @@ class TrainingApplicationAdminFormFactory
 		$this->addPaymentInfo($form);
 		$form->addSubmit('submit', 'Uložit');
 
-		$upcoming = $this->trainingDates->getPublicUpcoming();
+		$upcoming = $this->upcomingTrainingDates->getPublicUpcoming();
 		$dates = [];
 		if ($application->dateId) {
 			$dates[$application->dateId] = $this->trainingDates->formatDateVenueForAdmin($this->trainingDates->get($application->dateId));
