+++ site/app/Www/Presenters/HomepagePresenter.php
@@ -7,7 +7,7 @@ use MichalSpacekCz\Articles\Articles;
 use MichalSpacekCz\Interviews\Interviews;
 use MichalSpacekCz\Talks\Talks;
 use MichalSpacekCz\Training\CompanyTrainings;
-use MichalSpacekCz\Training\Dates\TrainingDates;
+use MichalSpacekCz\Training\Dates\UpcomingTrainingDates;
 use MichalSpacekCz\Training\Trainings;
 
 class HomepagePresenter extends BasePresenter
@@ -17,7 +17,7 @@ class HomepagePresenter extends BasePresenter
 		private readonly Articles $articles,
 		private readonly Interviews $interviews,
 		private readonly Talks $talks,
-		private readonly TrainingDates $trainingDates,
+		private readonly UpcomingTrainingDates $upcomingTrainingDates,
 		private readonly Trainings $trainings,
 		private readonly CompanyTrainings $companyTrainings,
 	) {
@@ -32,7 +32,7 @@ class HomepagePresenter extends BasePresenter
 		$this->template->talks = $this->talks->getAll(5);
 		$this->template->favoriteTalks = $this->talks->getFavorites();
 		$this->template->upcomingTalks = $this->talks->getUpcoming();
-		$this->template->upcomingTrainings = $this->trainingDates->getPublicUpcoming();
+		$this->template->upcomingTrainings = $this->upcomingTrainingDates->getPublicUpcoming();
 		$this->template->companyTrainings = $this->companyTrainings->getWithoutPublicUpcoming();
 		$this->template->interviews = $this->interviews->getAll(5);
 		$this->template->lastFreeSeats = $this->trainings->lastFreeSeatsAnyTraining($this->template->upcomingTrainings);
